split off data type specific tests + combine with dictionary tests
